need meta-description and meta keywords
Meta descriptions can be any length, but search engines truncate snippets longer than 160 characters. It is best to keep meta descriptions long enough that they're sufficiently descriptive but shorter than that 160-character limit.

Meta Keywords are a particular type of meta tag that appears in the HTML code of a Web page and helps tell search engines what the topic of the page is.